deserialized시 hash변수를 GetByte로 가져오는 것을 GetInt64로 정정

+++ SteamB23.Security/SecureInt64.cs
@@ -121,7 +121,7 @@ namespace System
         public SecureInt64(SerializationInfo info, StreamingContext context)
         {
             this.value = info.GetInt64("securedValue");
-            this.hash = info.GetByte("checksum");
+            this.hash = info.GetInt64("checksum");
         }
         #endregion
